Move existing Template exporter under :v1
