+++ app/controllers/dradis/plugins/projects/templates_controller.rb
@@ -5,7 +5,7 @@ module Dradis::Plugins::Projects
       export_manager_hash   = session[:export_manager].with_indifferent_access
       content_service_class = export_manager_hash[:content_service].constantize
 
-      exporter = Dradis::Plugins::Projects::Export::Template.new(
+      exporter = Dradis::Plugins::Projects::Export::V1::Template.new(
         content_service: content_service_class.new(plugin: Dradis::Plugins::Projects)
       )
